% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{Sum with NA Handling}
\usage{
plus(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Sum of \code{x} with \code{NA} handling.
}
\description{
Computes the sum of a numeric vector but returns \code{NA}
if all elements are missing.
}
\examples{
plus(c(1, 2, 3, NA))
plus(c(NA, NA, NA))
}
