% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_na.R
\name{sum_na}
\alias{sum_na}
\title{Sum with NA Handling}
\usage{
sum_na(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Sum of \code{x} with \code{NA} handling.
}
\description{
Computes the sum of a numeric vector but returns \code{NA}
if all elements are missing.
}
\examples{
sum_na(c(1, 2, 3, NA))
sum_na(c(NA, NA, NA))
}
